// backend/src/types/qatar/qatarFlightList.ts
type QatarFlightList = {
  flightNumber: string;
  departedDate: string;
  arrivalDate: string;
  operationType: string;
  segmentOfDeparture: string;
  segmetnOfArrival: string;
  colourSts: any;
  depPcs: number;
  std: string;
  sta: string;
  pieces: number;
  weight: number;
  volume: number;
  weightUnit: string;
  volumeUnit: string;
  flightStatus: string;
  bkgStatus: string;
  cargoTrackingMvtStausList: any;
};

export default QatarFlightList;